Extract MonadFresh & FreshT into its own module.
